+++ packages/cn-international-express-sdk/src/index.ts
@@ -3,3 +3,5 @@ export * from '@cn-international-express-sdk/yuntu';
 export * from '@cn-international-express-sdk/zhxt';
 export * from '@cn-international-express-sdk/yanwen';
 export * from '@cn-international-express-sdk/cainiao';
+export * from '@cn-international-express-sdk/jiehang';
+export * from '@cn-international-express-sdk/one7feia';