+++ packages/cn-international-express-sdk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # cn-international-express-sdk
 
+## 3.0.0
+
+### Major Changes
+
+- add jiehang-sdk
+- include one7feia-sdk
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @cn-international-express-sdk/jiehang@1.0.0
+
 ## 2.0.1
 
 ### Patch Changes